move harmful provider, filter out missing tests
